# Conflicts:
#	Assets/10.Effects/VFX/AddressableAssetsData/AssetGroups/Default Local Group.asset

+++ Assets/01.Scripts/Manager/SaveManager.cs
@@ -71,6 +71,10 @@ public static class SaveManager
         System.IO.DirectoryInfo di = new System.IO.DirectoryInfo(SAVE_PATH);
         foreach (System.IO.FileInfo File in di.GetFiles())
         {
+            if(File.Name == "ImportantGameData")
+            {
+                continue;
+            }
             Debug.Log(File.Name + "Á¦°Å");
             File.Delete();
         }
